pyproject.toml
[tool.poetry]
name = "auto-subtitles-generator"
version = "0.2.0"
authors = ["Gheorghe Andrei <gheorgheandrei13@gmail.com>"]
description = "Auto subtitles generator for videos"
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.26"
ffmpeg-python = "^0.2.0"
click = "^8.1.3"
python-dotenv = "^1.0.0"
httpx = {version = ">=0.22,<0.24", source = "pypi"}

[tool.poetry.group.dev.dependencies]
coverage = "^7.2.2"
flake8 = "^6.0.0"
poethepoet = "^0.19.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
auto_subtitles = "src.cli:auto_subtitles"

[tool.poe.tasks]
test = "python -m unittest test/"
coverage = "coverage run -m unittest test/"
lint = "flake8"
clean = """
# Multiline commands including comments work too. Unescaped whitespace is ignored.
rm -rf .coverage
       .mypy_cache
       .pytest_cache
       dist
       ./**/__pycache__
"""

[tool.coverage.run]
omit = [".*", "*/site-packages/*"]

[tool.flake8.run]
omit = [".*", "*/site-packages/*"]